+++ packages/api-evm/source/actions/index.ts
@@ -1,8 +1,8 @@
 export * from "./eth-block-number.js";
 export * from "./eth-call.js";
 export * from "./eth-get-balance.js";
-export * from "./get-code.js";
-export * from "./get-storage-at.js";
+export * from "./eth-get-code.js";
+export * from "./eth-get-storage-at.js";
 export * from "./net-listening.js";
 export * from "./net-peer-count.js";
 export * from "./web3-client-version.js";
